+++ src/de/mod.rs
@@ -86,20 +86,18 @@ pub struct Deserializer<'b> {
 }
 
 impl<'a> Deserializer<'a> {
-    fn new(slice: &'a [u8]) -> Deserializer<'_> {
+    /// Create a new `Deserializer`
+    pub fn new(slice: &'a [u8]) -> Deserializer<'_> {
         Deserializer { slice, index: 0 }
-    }
-
-    /// Return the current index in the buffer.
-    pub fn index(&self) -> usize {
-        self.index
     }
 
     fn eat_char(&mut self) {
         self.index += 1;
     }
 
-    fn end(&mut self) -> Result<usize> {
+    /// Check whether there is any unexpected data left in the buffer
+    /// and return the amount of data consumed
+    pub fn end(&mut self) -> Result<usize> {
         match self.parse_whitespace() {
             Some(_) => Err(Error::TrailingCharacters),
             None => Ok(self.index),